Rearranged GUI, Added extra Functions to spiralize
Rearranged the GUI so that the image is larger and appears on the left
side of the screen, while the controls are in a column on the right.

Fixed a bug that could occur with some functions regarding how floats
don't interact so cleanly with the modulo % operator in Python.

Added the ability to choose functions other than np.log to build the
tiling, so now you can play with similar mappings, esp. Mobius.